dev_prod_brgin

+++ src/main/java/com/srm/machinemonitor/Models/Other/CustomUserDetails.java
@@ -5,6 +5,7 @@ import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 
+import java.math.BigInteger;
 import java.util.Collection;
 import java.util.Collections;
 
@@ -31,7 +32,7 @@ public class CustomUserDetails implements UserDetails {
         return users.getUsername();
     }
 
-    public int getOrganizationId(){
+    public BigInteger getOrganizationId(){
         return users.getOrganizationId();
     }
