+++ src/main/java/com/srm/machinemonitor/Services/OrganizationDAO.java
@@ -4,13 +4,15 @@ import com.srm.machinemonitor.Models.Tables.Organizations;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface OrganizationDAO extends JpaRepository<Organizations, Integer> {
+import java.math.BigInteger;
+
+public interface OrganizationDAO extends JpaRepository<Organizations, BigInteger> {
     Organizations findByName(String name);
 
     @Query(value="SELECT id from organizations where name = ?1",nativeQuery = true)
-    Integer getIdByName(String name);
+    BigInteger getIdByName(String name);
 
-    Organizations getNameById(int id);
+    Organizations getNameById(BigInteger id);
 
     boolean existsByName(String organization);
 }
